# ARROW-4748: [Rust] [DataFusion] Optimize GROUP BY aggregate queries