Various optimizations to GROUP BY aggregate queries, resulting in a 25% improvement in the current benchmark, and close to a 2x improvement when running queries against large data sets.

- Re-use vector for grouping keys instead of creating a new vector for each row (benchmark went from 47 us to 39 us)
- Two pass approach where pass one builds a vector of references to accumulators, and the second pass processes one column at a time, updating the accumulators (now 37.6 us)
- Create aggregate schema once when plan is constructed, instead of once per batch per partition (now 36.5 us)
- Avoid converting final aggregate map into a vector (now 34.8 us)